move api method to a new file
