+++ server.js
@@ -7,23 +7,29 @@ const handle = app.getRequestHandler()
 
 const fetch = require('isomorphic-unfetch')
 
-const configs = require('./configs.js')
+// const configs = require('./configs.js')
+
+const fetchables = require('~/fetchables')
 
 app.prepare().then(() => {
     const server = express()
 
     server.get('/ig/posts', (req, res) => {
-        fetch( configs.ig.url + '/v1/users/' + configs.ig.userId + '/media/recent?access_token=' + configs.ig.token, {
-            method: 'GET',
-            credentials: 'include',
-            headers: { 'Content-Type': 'application/json' }
-        }).then( posts => {
-            return posts.json()
-        }).then( posts => {
-            res.send(posts)
-        }).catch( posts => {
-            res.json(posts)
+        const posts = fetchables.getInitalIgPosts()
+        return posts.resolution.then((res) => {
+           res.send(res)
         })
+        // fetch( configs.ig.url + '/v1/users/' + configs.ig.userId + '/media/recent?access_token=' + configs.ig.token, {
+        //     method: 'GET',
+        //     credentials: 'include',
+        //     headers: { 'Content-Type': 'application/json' }
+        // }).then( posts => {
+        //     return posts.json()
+        // }).then( posts => {
+        //     res.send(posts)
+        // }).catch( posts => {
+        //     res.json(posts)
+        // })
     })
 
     server.get('*', (req, res) => {
